Disable IRQ when sending mbox message
If AP did not read previous message yet, we must wait til it does.
The opreation of reading this state and sending the message has
to be atomic, since mbox_send can also be called from
mbox_irq_handler. Thus we disable IRQ handlers when reading this state
and writing mbox registers.

Signed-off-by: Marek Behún <marek.behun@nic.cz>