wtmi: assume constant frequency of Cortex-M3 clock
Currently the frequency of Cortex-M3 clock is being computed from the
register values. But every possible clock tree setting sets this clock
to 200 MHz, and the specification also says that this processor runs at
200 MHz.

Assume constant frequency of 200 MHz. We can get rid of clock_init() and
use constants in wait_ns() and udelay() functions.

Signed-off-by: Marek Behún <marek.behun@nic.cz>