wtmi: main: temporary gross hack to avoid colliding with TF-A output
When running as application for sys_init the sys_init code kicks the
TF-A on application processor before executing our firmware on the
secure coprocessor.

It seems that our printf() implementation is slow enough that when using
it, the console output gets mangled with initial console output from
TF-A.

Add a temporary gross hack (hopefully we will find a better solution in
the future) to solve this: it seems that using puts() / fputs() is
enough.

Signed-off-by: Marek Behún <marek.behun@nic.cz>
Suggested-by: Pali Rohár <pali@kernel.org>