wtmi: Don't program already programmed bits in efuse_raw_write()
In efuse_raw_write() we need first to read the OTP row, otherwise
writing won't work. We currently add the already written bits to the
value about to be written with bitwise OR. But if the row is masked,
reading will not return the real value of the row, but special value
0x0bad0acc0bad0acc. We can't add these bits to the value about to be
written.

Do not add bits returned from _efuse_read_row() to the values about to
be written.

Signed-off-by: Marek Behún <marek.behun@nic.cz>